Restructure to fix most tests
