+++ packages/squiggle-lang/src/public/SqValue/index.ts
@@ -1,7 +1,6 @@
 import { result } from "../../utility/result.js";
 import { SDate } from "../../utility/SDate.js";
 import {
-  toSimpleValue,
   Value,
   vCalculator,
   vDate,
@@ -9,7 +8,7 @@ import {
   vNumber,
   vString,
 } from "../../value/index.js";
-import { SimpleValue } from "../../value/simpleValue.js";
+import { SimpleValue, valueToSimpleValue } from "../../value/simpleValue.js";
 import { SqError } from "../SqError.js";
 import { SqValueContext } from "../SqValueContext.js";
 import { SqArray } from "./SqArray.js";
@@ -80,7 +79,7 @@ export abstract class SqAbstractValue<Type extends string, JSType> {
   }
 
   toSimple(): SimpleValue {
-    return toSimpleValue(this._value);
+    return valueToSimpleValue(this._value);
   }
 
   publicName() {
@@ -102,7 +101,7 @@ export class SqArrayValue extends SqAbstractValue<"Array", SimpleValue[]> {
   }
 
   asJS(): SimpleValue[] {
-    return this._value.value.map(toSimpleValue);
+    return this._value.value.map(valueToSimpleValue);
   }
 }
 
@@ -156,7 +155,7 @@ export class SqDistributionValue extends SqAbstractValue<"Dist", SimpleValue> {
   }
 
   asJS() {
-    return toSimpleValue(this._value);
+    return valueToSimpleValue(this._value);
   }
 }
 
@@ -172,7 +171,7 @@ export class SqLambdaValue extends SqAbstractValue<"Lambda", SimpleValue> {
   }
 
   asJS() {
-    return toSimpleValue(this._value);
+    return valueToSimpleValue(this._value);
   }
 
   toCalculator(): SqCalculatorValue | undefined {
@@ -207,7 +206,7 @@ export class SqDictValue extends SqAbstractValue<"Dict", SimpleValue> {
   }
 
   asJS(): SimpleValue {
-    return toSimpleValue(this._value);
+    return valueToSimpleValue(this._value);
   }
 }
 
@@ -255,7 +254,7 @@ export class SqPlotValue extends SqAbstractValue<"Plot", SimpleValue> {
   }
 
   asJS() {
-    return toSimpleValue(this._value);
+    return valueToSimpleValue(this._value);
   }
 }
 export class SqTableChartValue extends SqAbstractValue<
@@ -269,7 +268,7 @@ export class SqTableChartValue extends SqAbstractValue<
   }
 
   asJS() {
-    return toSimpleValue(this._value);
+    return valueToSimpleValue(this._value);
   }
 }
 export class SqCalculatorValue extends SqAbstractValue<
@@ -283,7 +282,7 @@ export class SqCalculatorValue extends SqAbstractValue<
   }
 
   asJS() {
-    return toSimpleValue(this._value);
+    return valueToSimpleValue(this._value);
   }
 
   override title() {
@@ -299,7 +298,7 @@ export class SqScaleValue extends SqAbstractValue<"Scale", SimpleValue> {
   }
 
   asJS() {
-    return toSimpleValue(this._value);
+    return valueToSimpleValue(this._value);
   }
 }
 
@@ -311,7 +310,7 @@ export class SqInputValue extends SqAbstractValue<"Input", SimpleValue> {
   }
 
   asJS() {
-    return toSimpleValue(this._value);
+    return valueToSimpleValue(this._value);
   }
 }
 
@@ -335,7 +334,7 @@ export class SqDomainValue extends SqAbstractValue<"Domain", SimpleValue> {
   }
 
   asJS() {
-    return toSimpleValue(this._value);
+    return valueToSimpleValue(this._value);
   }
 }
